Start work on upgrades UI, a compact font, plus probably a bunch of stuff im forgetting about

+++ Assets/Scripts/Upgrade.cs
@@ -23,7 +23,7 @@ public class Upgrade
 		this.growthPerUnit = growthPerUnit;
 	}
 
-	public double GetCurrentCost (int currentUnitCount)
+	public double GetCurrentCost (long currentUnitCount)
 	{
 		return baseCost * Math.Pow(costMultiplier, currentUnitCount);
 	}
